seeder_service/requirements.txt
mysql-connector==2.2.9
mysqlclient==1.3.12